requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.3.5
scipy>=1.7.3
scikit-learn>=1.0.2
joblib>=1.1.0

# Data processing
nltk>=3.7
tqdm>=4.62.3

# Visualization
matplotlib>=3.5.1
seaborn>=0.11.2
wordcloud>=1.8.1

# Web interface
flask>=2.0.2
flask-bootstrap>=3.3.7.1
werkzeug>=2.0.2

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development tools
black>=22.1.0
flake8>=4.0.1
isort>=5.10.1